requirements.txt
absl-py==2.2.2
accelerate==1.1.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2025.4.26
chardet==4.0.0
charset-normalizer==3.4.2
click==8.1.8
cmake==4.0.0
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
DataProperty==1.1.0
datasets==4.0.0
dbus-python==1.2.18
debugpy==1.8.13
decorator==5.2.1
defer==1.0.6
dill==0.3.8
distro==1.7.0
distro-info===1.1build1
docker-pycreds==0.4.0
docstring_parser==0.16
duplicity==0.8.21
einops==0.8.1
einops-exts==0.0.4
evaluate==0.4.0
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.14.1
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
future==0.18.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.8
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.31.0
idna==3.10
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==18.1.8
-e git+https://github.com/babylm/evaluation-pipeline-2024.git@d680c2875e2141d0db14f68c305caca9e6d86b85#egg=lm_eval
lockfile==0.12.2
louis==3.20.0
lxml==5.4.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
minicons==0.3.30
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==2.2.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
packaging==25.0
pandas==2.3.1
paramiko==2.9.3
parso==0.8.4
pathlib==1.0.1
pathvalidate==3.2.3
patsy==1.0.1
peft==0.15.2
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.3.8
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==3.12.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyarrow-hotfix==0.7
pybind11==2.13.6
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytablewriter==1.2.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
pytorch-ranger==0.1.1
pytz==2025.2
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
responses==0.18.0
rich==13.9.4
rouge_score==0.1.2
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentry-sdk==2.27.0
seqeval==1.2.2
setproctitle==1.3.6
shtab==1.7.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
speedtest-cli==2.1.3
sqlitedict==2.1.0
ssh-import-id==5.11
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
systemd-python==234
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.5.1+cu121
torch-optimizer==0.3.0
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.51.3
triton==3.1.0
typeguard==4.4.1
typepy==1.3.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tyro==0.9.2
tzdata==2025.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.4.0
usb-creator==0.3.7
wadllib==1.3.6
wandb==0.19.11
wcwidth==0.2.13
wonderwords==2.2.0
word2number==1.1
wrapt==1.17.2
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.20.0
zipp==1.0.0
zstandard==0.23.0
